+++ app/controllers/api/v1/open_library_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    class OpenLibraryController < ApplicationController
+      def show
+        @book = OpenLibraryService.new(params[:isbn]).book_info
+        render json: @book[:result], status: @book[:status?]
+      end
+
+    end
+  end
+end
