+++ app/models/user.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 class User < ApplicationRecord
   extend Devise::Models
   # Include default devise modules.
@@ -8,4 +6,5 @@ class User < ApplicationRecord
   include DeviseTokenAuth::Concerns::User
 
   has_many :rent
+
 end
