--- docs/api_docs/api_blueprint/api.apib
@@ -1,96 +1,0 @@
-Format: 1A
-Host: http://localhost:3000
-
-# Books API doc
-
-Documentation generated by Wolox - Fictium gem
-
-
-# Information and references
-
-API version: 1
-
-# Group Resources 
-
-# Books [/api/book]
-Handles api book.
-
-
-## List all available books [GET ]
-
-+ Response 200 (application/json)
- 
-  + Header
-
-            access-token: 52yI7kKxAN6O_QMFULhumg
-            token-type: Bearer
-            client: b0RalN6TEGCYWWD-T_LD7g
-            expiry: 1625598757
-            uid: merle@jerde.io
- 
-  + Body
-
-      ```
-      {"page":[{"author":"Ms. Darin Wilderman","title":"The Millstone","genre":"Short story","editor":"J. M. Dent","year":"1001","image":"https://loremflickr.com/50/60","slug":"the-millstone"},{"author":"Rickey Labadie Jr.","title":"Sleep the Brave","genre":"Realistic fiction","editor":"Blackstaff Press","year":"2009","image":"https://loremflickr.com/50/60","slug":"sleep-the-brave"},{"author":"Napoleon Cormier","title":"Cabbages and Kings","genre":"Short story","editor":"Virgin Publishing","year":"1248","image":"https://loremflickr.com/50/60","slug":"cabbages-and-kings"},{"author":"Norene Tremblay","title":"I Will Fear No Evil","genre":"Fiction narrative","editor":"Tammi","year":"1159","image":"https://loremflickr.com/50/60","slug":"i-will-fear-no-evil"},{"author":"Aubrey Collins","title":"The Curious Incident of the Dog in the Night-Time","genre":"Fantasy","editor":"Carlton Books","year":"1286","image":"https://loremflickr.com/50/60","slug":"the-curious-incident-of-the-dog-in-the-night-time"}],"count":5,"total_pages":1,"total_count":5,"current_page":1,"previous_page":null,"next_page":null,"next_page_url":null,"previous_page_url":null}
-      ```
-
-
-+ Response 200 (application/json)
- 
-  + Header
-            access-token: ei227KTsUYSZmFXrObs_oQ
-            token-type: Bearer
-            client: Gx0o9G3Hr2myLBpXyVAuMg
-            expiry: 1625598758
-            uid: jamika.ebert@witting.biz
-  + Body
-
-      ```
-      {"page":[{"author":"Emmanuel Conroy","title":"The Little Foxes","genre":"Mythology","editor":"Harry N. Abrams, Inc.","year":"1859","image":"https://loremflickr.com/50/60","slug":"the-little-foxes"}],"count":1,"total_pages":1,"total_count":1,"current_page":1,"previous_page":null,"next_page":null,"next_page_url":null,"previous_page_url":null}
-      ```
-
-
-## Show details of a book. [GET ]
-
-+ Response 200 (application/json)
- 
-  + Header
-            access-token: dJpMCoUI2uy4NKeTShh2_A
-            token-type: Bearer
-            client: 2tN-pkXzetSEiNag9brZGg
-            expiry: 1625598759
-            uid: eusebio_white@stroman.name
- 
-  + Body
-
-      ```
-      {"author":"Mitchell Fay II","title":"This Side of Paradise","genre":"Mystery","editor":"University of Nebraska Press","year":"1369","image":"https://loremflickr.com/50/60","slug":"this-side-of-paradise"}
-      ```
-
-
-+ Response 404 (application/json)
- 
-  + Body
-
-# Open libraries [/api/open_library]
-Handles api open_library.
-
-
-## Show details of an open library. [GET ]
-
-+ Response 200 (application/json)
- 
-  + Body
-
-      ```
-      {"ISBN":"2","title":"Pragmatic Realism, Religious Truth, and Antitheodicy","subtitle":"On Viewing the World by Acknowledging the Other","number_of_pages":223,"authors":[{"url":"https://openlibrary.org/authors/OL5422937A/Sami_Pihlström","name":"Sami Pihlström"}]}
-      ```
-
-
-+ Response 404 (application/json)
- 
-  + Body
-
-      ```
-      {"status":404,"errors":[{"code":"001","message":"Book not found"},{"code":"002","message":"ISBN cannot be alphabetic"}]}
-      ```
