--- docs/api_docs/open_api/3.0.0/swagger.json
@@ -1,134 +1,0 @@
-{
-  "openapi": "3.0.0",
-  "info": {
-    "title": "Books API doc",
-    "description": "Documentation generated by Wolox - Fictium gem",
-    "version": "1"
-  },
-  "paths": {
-    "/api/book": {
-      "description": "",
-      "get": {
-        "tags": [
-
-        ],
-        "description": "Show details of a book.",
-        "parameters": [
-
-        ],
-        "responses": {
-          "default": {
-            "description": "with valid value {}",
-            "content": {
-              "application/json": {
-                "example": "{\"author\":\"Mitchell Fay II\",\"title\":\"This Side of Paradise\",\"genre\":\"Mystery\",\"editor\":\"University of Nebraska Press\",\"year\":\"1369\",\"image\":\"https://loremflickr.com/50/60\",\"slug\":\"this-side-of-paradise\"}"
-              }
-            }
-          },
-          "404": {
-            "description": "with invalid value {}",
-            "content": {
-              "application/json": {
-                "example": "{\"error\":\"can't find record with friendly id: \\\"ullam_doloremque\\\"\"}"
-              }
-            }
-          }
-        },
-        "deprecated": false
-      }
-    },
-    "/api/book_suggestion": {
-      "description": "",
-      "post": {
-        "tags": [
-
-        ],
-        "description": "Create a new book suggestion.",
-        "parameters": [
-
-        ],
-        "requestBody": {
-          "content": {
-            "application/x-www-form-urlencoded": {
-              "example": "author=Rev.+Jennell+Jakubowski&editor=Carlton+Publishing+Group&link=http%3A%2F%2Fhermann.co%2Frusty&title=Pale+Kings+and+Princes&year=1409"
-            }
-          }
-        },
-        "responses": {
-          "default": {
-            "description": "alone with valid parameters required {}",
-            "content": {
-              "application/json": {
-                "example": "{\"message\":\"Book Suggestion created\"}"
-              }
-            }
-          },
-          "201": {
-            "description": "with valid synopsis param {}",
-            "content": {
-              "application/json": {
-                "example": "{\"message\":\"Book Suggestion created\"}"
-              }
-            }
-          },
-          "422": {
-            "description": "without parameters required {}",
-            "content": {
-              "application/json": {
-                "example": "{\"error\":{\"title\":[\"can't be blank\"],\"editor\":[\"can't be blank\"],\"link\":[\"is invalid\",\"can't be blank\"],\"year\":[\"is not a number\",\"is too short (minimum is 1 character)\",\"can't be blank\"],\"author\":[\"is invalid\",\"can't be blank\"]}}"
-              }
-            }
-          },
-          "400": {
-            "description": "with invalid param {}",
-            "content": {
-              "application/json": {
-                "example": "{\"error\":\"found unpermitted parameter: :lorem\"}"
-              }
-            }
-          },
-          "404": {
-            "description": "with invalid param user_id {}",
-            "content": {
-              "application/json": {
-                "example": "{\"error\":\"Couldn't find User with 'id'=a\"}"
-              }
-            }
-          }
-        },
-        "deprecated": false
-      }
-    },
-    "/api/open_library": {
-      "description": "",
-      "get": {
-        "tags": [
-
-        ],
-        "description": "Show details of an open library.",
-        "parameters": [
-
-        ],
-        "responses": {
-          "default": {
-            "description": "with valid value {}",
-            "content": {
-              "application/json": {
-                "example": "{\"ISBN\":\"2\",\"title\":\"Pragmatic Realism, Religious Truth, and Antitheodicy\",\"subtitle\":\"On Viewing the World by Acknowledging the Other\",\"number_of_pages\":223,\"authors\":[{\"url\":\"https://openlibrary.org/authors/OL5422937A/Sami_Pihlström\",\"name\":\"Sami Pihlström\"}]}"
-              }
-            }
-          },
-          "404": {
-            "description": "with invalid value {}",
-            "content": {
-              "application/json": {
-                "example": "{\"status\":404,\"errors\":[{\"code\":\"001\",\"message\":\"Book not found\"},{\"code\":\"002\",\"message\":\"ISBN cannot be alphabetic\"}]}"
-              }
-            }
-          }
-        },
-        "deprecated": false
-      }
-    }
-  }
-}
